requirements.txt
amqp==2.6.0
asgiref==3.2.7
astroid==2.4.0
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
cheroot==8.3.0
CherryPy==18.6.0
colorama==0.4.3
Django==3.0.4
django-filter==2.3.0
django-fitbit==0.3.0
djangorestframework==3.11.0
fitbit==0.3.1
future==0.18.2
idna==2.9
isort==4.3.21
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.1
jaraco.text==3.2.0
kombu==4.6.11
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
more-itertools==8.4.0
numpy==1.19.0
oauthlib==3.1.0
pandas==1.0.5
portend==2.6
pylint==2.5.0
python-dateutil==2.8.1
python-oauth2==1.1.1
pytz==2019.3
pywin32==228
requests==2.24.0
requests-oauthlib==1.3.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
tempora==3.0.0
toml==0.10.0
urllib3==1.25.9
vine==1.3.0
wrapt==1.12.1
zc.lockfile==2.0
